Add runConsole subcommand
This command executes a single assembly language program, then emits
a JSON structure containng the program stdout, final register state,
and state of the first 1024 words of its data segment.

This subcommand is suitable for interactive use from a console,
without the use a web browser (so it differs from runELF) but also
provides detailed output of the final state of the emulator (so
it differs from runBatch).